frameworkify schema tests
