+++ test/common/json/BUILD
@@ -2,11 +2,18 @@ package(default_visibility = ["//visibility:public"])
 
 load("//bazel:envoy_build_system.bzl", "envoy_cc_test")
 
+filegroup(
+    name = "schema_test_generator",
+    srcs = glob([
+        "test_data/*.py",
+    ]),
+)
+
 envoy_cc_test(
     name = "config_schemas_test",
     srcs = ["config_schemas_test.cc"],
     data = [
-        "test_data/generate_test_data.py",
+        "//test/common/json:schema_test_generator",
     ],
     deps = [
         "//source/common/json:config_schemas_lib",
