+++ test/common/json/config_schemas_test.cc
@@ -13,7 +13,7 @@ std::vector<std::string> generateTestInputs() {
   TestEnvironment::exec(
       {TestEnvironment::runfilesPath("test/common/json/test_data/generate_test_data.py")});
 
-  std::string test_path = TestEnvironment::temporaryDirectory();
+  std::string test_path = TestEnvironment::temporaryDirectory() + "/config_schema_test";
   DIR* dir = opendir(test_path.c_str());
   if (!dir) {
     throw std::runtime_error("Generated config schema test directory not found");
